Enforce tags to be lowercase
